Edited EOL spec helper to truncate only tables that need truncating, making it slightly faster
git-svn-id: https://repository.eol.org/eol/trunk@1785 78829999-583a-0410-bd01-8a9b849fd409

+++ spec/eol_spec_helpers.rb
@@ -75,7 +75,8 @@ module EOL
           conn.tables.each   do |table|
             unless table == 'schema_migrations'
               puts "[#{conn.instance_eval { @config[:database] }}].`#{table}`" if options[:verbose]
-              conn.execute "TRUNCATE TABLE`#{table}`"
+              count_rows = conn.execute("SELECT 1 FROM #{table} LIMIT 1")
+              conn.execute "TRUNCATE TABLE`#{table}`" if count_rows.num_rows > 0
             end
           end
         end
